Delete schedule working - identified bug when deleting scheduled bb

+++ Billboard/src/Clients/ControlPanel/ControlPanelTools/checkDurationVsRecurrence.java
@@ -3,13 +3,11 @@ package Clients.ControlPanel.ControlPanelTools;
 import javax.swing.*;
 
 import static Clients.ControlPanel.ControlPanelInterface.ControlPanelInterface.schedulePanel;
-import static Clients.ControlPanel.ControlPanelTools.DateChooser.minRec;
-import static Clients.ControlPanel.ControlPanelTools.DurationSetter.duration;
 import static Clients.ControlPanel.ControlPanelTools.DurationSetter.tf_duration;
 
 public class checkDurationVsRecurrence {
 
-    public static void  checkDVsR (int minRec, long duration){
+    public static void  checkDVsR (int minRec, int duration) {
 
         if (duration > minRec) {
             JOptionPane.showMessageDialog(schedulePanel,
@@ -17,7 +15,6 @@ public class checkDurationVsRecurrence {
                     "Input Error", JOptionPane.ERROR_MESSAGE);
             tf_duration.setText("");
         }
-
     }
 
 }
